+++ plugin/src/main/java/net/okocraft/box/plugin/database/table/PlayerTable.java
@@ -3,7 +3,7 @@ package net.okocraft.box.plugin.database.table;
 import net.okocraft.box.plugin.Box;
 import net.okocraft.box.plugin.database.connector.Database;
 import net.okocraft.box.plugin.model.User;
-import net.okocraft.box.result.UserCheckResult;
+import net.okocraft.box.plugin.result.UserCheckResult;
 import org.jetbrains.annotations.NotNull;
 
 import java.sql.Connection;
